Separated T2Analysis into a separate file
